requirements.txt
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.4
django-cors-headers==2.1.0
djangorestframework==3.6.3
djangorestframework-jwt==1.10.0
gunicorn==19.7.1
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.1
PyJWT==1.5.2
pytz==2017.2
urllib3==1.13.1
vine==1.1.3
whitenoise==3.3.0
